--- JeongSungwon/index.js
@@ -1,25 +1,0 @@
-import express from "express"
-const app = express();
-
-const PORT = 4000
-
-const handleListening = () => console.log(`Listening on: http://localhost:${PORT}`);
-
-
-const handleHome = (req, res) => res.send('Hello from my ass');
-
-
-const handleProfile = (req, res) => res.send("You are on my profile");
-
-const betweenHome = (req, res, next) => {
-  console.log('Between');
-  next();
-}
-
-app.use(betweenHome);
-
-app.get("/", betweenHome, handleHome);
-
-app.get("/profile", handleProfile)
-
-app.listen(PORT, handleListening);
